+++ NoteWithFireStore/NoteWithFireStore/Utils/Alertable.swift
@@ -50,8 +50,5 @@ public extension Alertable where Self: UIViewController {
          ))
          self.present(alert, animated: true, completion: nil)
     }
-
-    
-
 }
